backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
stripe
openai
pdfminer.six
python-docx
requests
beautifulsoup4
